NS-100 Move CalNet load up into cache_utils for easier relocation
